dtprobed: start as early as possible
We'd like to start dtprobed *before* multi-user.target so that the
majority of daemons (those started after that point, in parallel)
already find it running and can register any USDT probes they may
contain.

We also need to "systemctl enable" everything at install time, since
dtprobed et al don't appear in the presets, so start disabled by
default.  (In the specfile, we not only need to do that but prod
udev to get the udev rules loaded.  This is peculiar since udev
uses filesystem notifies to do that itself, but if that's in the
packaging standard, so be it... we even need to start the thing
by hand after installation, because the systemd macros rely on
dtprobed being named in the presets, which it is not.)

Also turn off RestrictSUIDSGID: it's a good idea but it relies
on a seccomp filter, which stops us from enabling strict seccomp
later on for dtprobed's parser child.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>