Simplify strcmp.S
The code branched depending on whether the first string is known
to be a pointer to D-managed storage.  If we simply assume that
the storage is never D-managed, we can save two instructions as
well as a conditional jump (the bane of the BPF verifier).  The
execution path ends up a couple of instructions either longer or
shorter depending on the case.  The resulting implementation is
closer to that of other string subroutines.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>