Add a CPC provider
Use libpfm to populate CPC probes.

The PC is placed into arg0 (if kernel) or arg1 (if user-space),
the other arg being 0.  This initialization is in a new function
dt_cg_tramp_copy_PC_from_regs() so that it can be used by any other
provider (like profile) that needs it.

Current limitations include:

*)  no support for optional attributes

*)  no support for raw events

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>