cmd, libdtrace: fix potentially unsafe printf-like calls
GCC 12 flagged some calls to printf-style functions for using a dynamic
string as format without arguments.  If the dymanic string somehow can
be made to contain a substitution sequence (%<something>) this would be
unsafe.

Due to some special case handing for error strings ending in a newline
character, the default error, liberr, and drop strings are adjusted as
well.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>