Increase the maximum size of strstab
The strtab size was limited in the legacy version to 0x7fff.  This
was generally sufficient because of how the strtab was implemented.
Now that all strings used in a probe program are stored in a single
strtab (incl. probe description names), a higher limit is needed.

All string offsets (and strtab size calculations) are using ssize_t,
so we use SSIZE_MAX as the upper limit.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>