proc: use pidfd_open(..., PIDFD_THREAD) if available
With luck this approach might possibly perhaps go upstream!

We still fall back to waitfd if using PIDFD_THREAD doesn't work.
Complicated somewhat by having to use SYS_syscall() if a glibc too old
to support pidfd_open() isn't available (i.e. all released ones: it'll
be in 2.36), and by the likelihood that PIDFD_THREAD isn't in the
header even if the header is present, but otherwise fairly trivial.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>