Fix profile-provider probe args arg0 and arg1
Set arg0 and arg1 for the profile provider.

Note that arg2 is still not unimplemented.

Add stronger tests for profile-* probe args.  (Since the new tests use
a target process that might be running on a different CPU from where
a tick-* probe fires, we do not add corresponding tick-* tests.
On the other hand, testing arg0 and arg1 for profile-* ought to be
sufficient.)

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>