Change condition to bypass trampoline generation
Originally, all providers were backed by system level probes and needed
a trampoline to be generated to convert from a BPF context to the DTrace
context DTrace expects.  With the introduction of pid probes, we gained
a provider that did not use a trampoline because it is implemented by
adding clauses to an underlying probe (using the same DTrace context as
the underlying probe).

But in a more generic sense (and soon to be used for SDT providers), we
eed to be able to support providers that do not generate programs that
are directly attached to system level probes while still needing to be
able to generate a trampoline.  So, the existance of a trampoline hook
is not sufficient.

Fortunately, all providers specify a BPF prgoram type that indicates the
system level probe type the program need to be generated for.  For
providers that do not generate actual programs, we can therefore specify
BPF_PROG_TYPE_UNSPEC as program type, and test for that to determine
whether the trampoline function is to be called at program load time.

Future patches will introduce a code path where a trampoline function is
called for a non-program load case.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>