probe: ensure probe argument info is retrieved only once per probe
Throughout the compilation process there are multiple points where
probe argument information is requested from the providers.  This
data cannot change during a runtime of DTrace and there is therefore
no reason to go through the process of retrieving it more than once.

The prp->argc (probe argument count) is used to determine if argument
data has been obtained.  It is initially -1, and will be 0 or greater
once argument data has been requested.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>