cg: support save/restore of probe arguments
With the upcoming introduction of dependent probes (probes that use
another underlying probe as trigger), we need to be able to save the
probe arguments for the underlying probe so that the arguments can be
modified in preparation for execution of the clauses of the dependent
probe.  After all such clauses have been executed, the probe arguments
need to be restored in case there is another dependent probe (with
possibly different arguments).

Dependent probes can only be one layer deep so we only eveyr need one
copy of the arguments.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>