htab: support lookups for buckets with multiple entries
Probes are represented by a quadruple naming scheme.  For easy lookup,
hashtables are kept based on each naming component.  When performing
lookups based on one of those hashtables, the head element was always
returned which is obviously wrong.

This patch introduced dt_htab_find() which provides for lookups using
a template entry and a secondary comparison function that is used to
evaluate the different entries in the bucket that match the template.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>