bpf: add 'agggen' BPF map
The 'agggen' BPF map associates a generation counter with an aggregation
ID.  It will be used to determine whether the aggregation data in the
'aggs' map for aggregations associated with a given aggregation ID is
valid or stale.

The initial generation counter for each aggregation is 1 (because 0 is
used to indicate a no-data condition).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>