Restore order of bcopy() arg evaluations
In c26e2e5f279a ("Add support for copyinto() subroutine"), the order
of arg evaluations for bcopy() was changed from evaluating the last
arg (size) last to evaluating it first.  This has semantic implications
when the order of arguments matters.

Restore the previous order (to agree with Solaris semantics) and add
tests for both bcopy() and copyinto().

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>