Implement clear()
The clear() action is implemented by incrementing the global generation
counter for an aggregation name and clearing the data for matching
aggregations within the consumer.  If aggregation data in a snapshot is
found to have an older generation, it is discarded.  The producer side
will already clear its data when the generation counter is found to be
higher than the generation of the data in a CPU's agg buffer.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>