Implement switchrate-based consumer loop
The switchrate option was not implemented in the BPF-based DTrace
implementation.  And commit 96d4dfeb (" Eliminate obsolete dt_last*")
removed the dtp->dt_lastswitch member from struct dtrace_hdl.  This
patch brings dt_lastswitch back and implements the switchrate option
as a way to limit the frequency of buffer polls.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>