tests: Use distinct trigger for copyin* tests
Many copyin* tests implement a trigger by running system("echo foo")
in the BEGIN probe.  This requires a destructive action, which is okay
but does make these "unit tests" rely on features unrelated to the
tests.

The main problem, however, is that the check that the corresponding
write:entry probe has fired on the correct process becomes very weak
-- e.g., just checking that a particular char is '-'.  And in the case
of copyinto, this check is unnecessarily complicated.  Finally, if
such a process is not found, the test would hang.

Use @@trigger.  E.g., delaydie will quickly terminate, emiting a simple
message, which is all we need for these tests.

The existing predicates check being able to:
- cast the return value of copyin() to string
- use indexing on the cast value
- use indexing on the return value of copyinstr()
Each such predicate is replaced with a conditionalized printf.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>