test: fix speculation/tst.SingleCPU.d
This test is meant to verify that simultaneously-active speculations get
drained and committed/discarded appropriately without losing any of
them.  Unfortunately, it never sets nspec, which is 1 by default, so it
actually runs out of speculations almost immediately: we need at least
16 for this test, which is not so coincidentally the maximum this
implementation supports right now.  Boosting that reveals another
problem: we print out the speculation ID, which is not in any sense a
guaranteed value, so we get lots of test failures: what we *actually*
care about is that the n value printed is suitable (indicating that the
first speculation in any block of (64 / 4) was committed and the rest
were discarded, and who cares what its ID is).

We also knock the switchrate down again because if we knock it up any
higher the discards (which are processed in userspace) don't get
processed fast enough to keep up, and we run out.  We consume 16
speculations in 6.4ms, so set the switchrate to 3ms to give us a bit of
headroom.  (This will still be racy on highly-loaded test systems where
the polls just don't happen fast enough because of system load: if that
happens, we can tag the test suitably in future.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>