Add support for uregs[]
The uregs[] essentially access (struct pt_regs) for the current task.
Therefore, we use the combination of BPF helper functions
bpf_task_pt_regs(bpf_get_current_task_btf()) to access pt_regs.

Note that older kernels might not support these helper functions.
So dt_bpf_init_helpers() checks, and if the needed functions are
missing an error results.

One place to put uregs[] support is in bpf/get_bvar.c.  On the other
hand, that means that the build system needs a gcc-bpf-unknown-none
that knows about these functions.  So, just generate the code ourselves
to call the two helper functions;  this is not much more difficult
than calling dt_get_bvar().

Further note that the documented symbolic indices to uregs[] seem
incorrect.  For x86, R_UESP is documented but not defined, and there
are many more indices that are defined but not documented.  For ARM,
the documented indices bear little resemblance to the defined (and
sensible!) values.

The support for symbolic indices to uregs[] on x86 mimics that found
in the legacy DTrace port to Linux -- that is, dtrace_getreg() from
kernel code arch/x86/dtrace/dtrace_isa_x86_64.c .  The logic is
convoluted and possibly incorrect.  Some uregs[] indices do not read
from (struct pt_regs) but from task->thread.*, complicating the
implementation.  This matter warrants further attention, but for now
mimicking the legacy port seems sufficient.

For the time being, there is no annotating disassembly with "uregs".
This can be added back in a future patch once other dust has settled.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>