Use highest bit to decide if PC is user space or kernel
Currently, we decide if a PC is user space or kernel based on whether
the BPF helper function probe_read_kernel() can read the address.
On systems where this BPF helper function does not exist, there is a
back-up mechanism that simply checks the highest bit of the PC.

Apparently, on some systems, the helper function can read addresses
even in user space.  That is, the mechanism that uses the helper function
is not reliable.

So, just use the "back-up mechanism" all the time.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>