test: Fix scripting test for "quiet"
This test was broken because the runtest.sh test harness was ignoring
the initial #! line, invoking the test as a D script rather than as an
interpreter file.  Hence, the -q was being ignored;  the .r results
file fit the actual (incorrect) behavior rather than the anticipated
(correct) behavior.

Use interpreter-file support in runtest.sh to launch the test as an
interpreter file.  Fix the results file to match expected "quiet"
results.  Rename the file from "quite" to "quiet."

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>