test: Switch kernel pointer in tracemem() test
The tracemem() test dumped a portion of kernel memory, comparing
different tracemem() output for the same kernel memory for consistency.
Since it was looking at task_struct for the current thread, however,
the underlying data could conceivably change as the test ran, leading
to inconsistency and a false negative.

Change the kernel memory to be the first 14 bytes of `linux_banner,
which should remain "Linux version ".

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>