provider: change the trampoline return type
The trampoline callback can now return 1 to indicate that no calls
should be made to clause functions (or 0 to retain the regular
operation of calling the clause functions).

This will be used by some providers to attach to probes that perform
auxiliary operations, e.g. setting up some data items that will be
needed for the actual probe processing.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>