test: Fix specsize tests
It used to be that specsize specified the size of some speculative
buffers.

In the current implementation, the producer writes all speculative
data to the principal buffer and the consumer reads the data, placing
it into growable linked lists until commit/discard disposition is
decided.  So, an extremely high specsize does not necessarily cause
a problem:  no extremely high allocation necessarily occurs.

Modify the tests accordingly.  Update .r results files for any gratuitous
formatting changes in the new implementation.  Ensure there is a .r
results file for each test.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>