uprobes, libproc: handle uprobes to procs in different fs namespaces
The specification for a uprobe includes a *mapping name*, which is
documented as the "path to [the] executable or library" where the probe
is placed.  Unstated in the documentation is what filesystem namespace
this path is relative to, because -- obviously enough -- it's relative
to the task placing the uprobe.

Unfortunately we get our mapping names, ultimately, from the names in
/proc/$pid/maps, and those names are relative to the fs namespace of the
process being queried.  So we cannot just use that name unmodified: the
file might not be there, or indeed anywhere accessible to the
prober. Worse yet, if we're really unlucky some *other* process might be
there and will have a probe stuffed in it at a thoroughly unsuitable
location. One can imagine an attacker setting up a mount namespace with
a /bin/su of their own in it, and putting a DTrace probe in it such that
we stick uprobes in the *real* /bin/su and oh dear. This can probably
only cause a crash, but it still feels very wrong.

Worse yet, when started by systemd dtprobed runs in a non-default
filesystem namespace in which /home is inaccessible: put all this
together and you can't create DTrace probes in binaries you compile
yourself under /home, unless you restart dtprobed yourself outside
systemd's aegis (or hack the service file to turn off ProtectHome).

So fix this by using the map names from /proc/$pid/map_files, which we
already acquire for the purposes of opening the ELF executable in
invasive mode.  For that, we can fall back to the old GETMAPFD approach,
but given that that is an out-of-tree patch and was dead, and we're not
even looking for an fd here but rather a filename, and because the
conseqeuences of just using a filename out of /proc/$pid/maps are so
bad, we'll simply not implement a fallback here.  USDT requires
map_files built into the kernel and that's all.

This makes uprobe_events less informative (no filenames any more), but
so be it. We can always add a filename as another parameter or something
if we want to.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>