usdt: provide probes properly when >1 is specified
dt_pid_create_usdt_probes is called by dt_pid_create_probes, which
despite its name is usually called by the compiler to set the context
for one probe at a time, not all at once. So dt_pid_create_usdt_probes
is usually (currently: always) called in a mode in which it scans all
the probes and then creates only those it has been asked to create.
So we really do need to not skip invocations after the first.

This will obviously not work well for wildcard probe handling, but for
now the inefficient approach of just repeated rescanning should work
well enough.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>