parser: check for all types of char
A while back I changed the check for IS_CHAR() in the parser to check
specifically for a char in the native signedness.  This can fail now
that the kernel might be using a char in the non-native signedness.

Consider all such types to be equally-valid chars.  We really
don't care if they're signed or not, they're still chars.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>