Consolidate tvar/assoc pointer access
To load or store a thread-local variable or associative array
element, one must generate a dynamic pointer and then access the
pointer.  Currently, there are four code paths (load or store,
tvar or assoc).  However, generating the dynamic pointer depends
only on whether it's a tvar or assoc, not on whether we will load
or store.  Further, accessing the pointer depends only on whether
we load or store, not on whether the pointer represents a tvar or
assoc.  So, there is an opportunity to consolidate code.

In this patch, consolidate the pointer-access code.  In the next
patch, we will consolidate the pointer generation.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>