spec: correct semantics of specsize overflow
The speculations implementation (commit 7fe6ce9a "consume, cg: implement
speculations") did not detect speculation data exceeding the specsize
limit until the consumer added data fragments to a speculation buffer.
As a result, all statements in clauses were always executed.

The legacy behaviour is to abort clause execution once a data recording
action causes the speculation buffer to overflow.  That causes any
further statements in the clause to be skipped.  The testsuite contains
an explicit test for this: tst.SpecSizeVariations4.d.  Unfortunately,
the test (and expected results) were amended to match the (incorrect)
new semantics.

This patch restores the legacy behaviour by tracking the size of the
speculation across clauses that contribute to a given speculation, so
that we can detect speculation buffer overflows as they occur due to
a specific data recording action.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>